refactor(analysis): use UnifiedParser in MetricThresholdChecker
MetricThresholdChecker was duplicating UnifiedParser's scanning logic,
requiring manual AttributeDescriptor registration which was bug-prone.

Changes:
- Add UnifiedParser.parse() as public API for library usage
- Refactor MetricThresholdChecker to use UnifiedParser.parse()
- Remove duplicated ProjectScanner setup code
- AttributeDescriptors now included automatically

This eliminates code duplication and ensures MetricThresholdChecker
benefits from any UnifiedParser improvements.
